Allow fractional movement costs

+++ common/unittype.h
@@ -478,7 +478,7 @@ struct unit_type {
                      engineers)*/
   int attack_strength;
   int defense_strength;
-  int move_rate;
+  float move_rate;
   int unknown_move_cost; // See utype_unknown_move_cost().
 
   struct advance *require_advance; // may be nullptr
